requirements.txt
matplotlib==3.7.2
numpy==2.2.1
Pillow==10.0.1
torchsummary==1.5.1
tqdm==4.66.4
